Move Header to its own component with own styles
