+++ components/Layout.js
@@ -1,0 +1,12 @@
+import Header from "./Header";
+
+const Layout = ({ children }) => {
+  return (
+    <>
+      <Header />
+      <main>{children}</main>
+    </>
+  );
+};
+
+export default Layout;
