+++ styles/Header.module.css
@@ -1,0 +1,31 @@
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1rem 2rem;
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+}
+
+.logo {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.appName {
+  margin-left: 12px;
+  font-weight: bold;
+  color: #335bf1;
+}
+
+.left,
+.right {
+  display: flex;
+  align-items: center;
+}
+
+.right > :not(:first-child) {
+  margin-left: 1rem;
+}
