+++ styles/Home.module.css
@@ -7,27 +7,6 @@
   align-items: center;
 }
 
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 2rem;
-  width: 100%;
-  height: 100%;
-  margin: 0 auto;
-}
-
-.nav {
-  float: right;
-  font-size: 0.875rem;
-}
-
-.nav a {
-  color: #335bf1;
-  font-weight: 600;
-  padding: 0.75em 1.5em;
-}
-
 .main {
   max-width: 540px;
   padding: 5rem 0;
@@ -181,16 +160,7 @@
 }
 
 .logo {
-  display: flex;
-  justify-content: center;
-  align-items: center;
   height: 1.5rem;
-}
-
-.appName {
-  margin-left: 12px;
-  font-weight: bold;
-  color: #335bf1;
 }
 
 .links {
@@ -212,14 +182,4 @@
 
 .linkText:hover {
   border-bottom: 2px solid #335bf1;
-}
-
-.left,
-.right {
-  display: flex;
-  align-items: center;
-}
-
-.right > :not(:first-child) {
-  margin-left: 1rem;
 }
